# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ORCID Link Service",
        "description": "The *ORCID Link Service* provides an API to enable the linking of a KBase\n user account to an ORCID account. This \"link\" consists of a [Link Record](#user-content-header_type_linkrecord) which \n contains a KBase username, ORCID id, ORCID access token, and a few other fields. This link record allows\n KBase to create tools and services which utilize the ORCID api to view or modify\n certain aspects of a users ORCID profile.\n\nOnce connected, *ORCID Link* enables certain integrations, including:\n\n- syncing your KBase profile from your ORCID profile\n- creating and managing KBase public Narratives within your ORCID profile",
        "termsOfService": "https://www.kbase.us/about/terms-and-conditions-v2/",
        "contact": {
            "name": "KBase, Lawrence Berkeley National Laboratory, DOE",
            "url": "https://www.kbase.us/",
            "email": "engage@kbase.us"
        },
        "license": {
            "name": "The MIT License",
            "url": "https://github.com/kbase/kb_sdk/blob/develop/LICENSE.md"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/linking-sessions": {
            "post": {
                "tags": [
                    "linking-sessions"
                ],
                "summary": "Create Linking Session",
                "description": "Create Linking Session\n\nCreates a new \"linking session\"; resulting in a linking session created in the database,\nand the id for it returned for usage in an interactive linking session.",
                "operationId": "create_linking_session_linking_sessions_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The linking session has been successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateLinkingSessionResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/docs": {
            "get": {
                "tags": [
                    "misc"
                ],
                "summary": "Docs",
                "description": "Get API Documentation\n\nProvides a web interface to the auto-generated API docs.",
                "operationId": "docs_docs_get",
                "responses": {
                    "200": {
                        "description": "Successfully returned the api docs",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateLinkingSessionResult": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "maxLength": 36,
                        "minLength": 36,
                        "title": "Session Id",
                        "description": "The linking session id"
                    }
                },
                "type": "object",
                "required": [
                    "session_id"
                ],
                "title": "CreateLinkingSessionResult"
            },
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string",
                        "maxLength": 25,
                        "minLength": 5,
                        "title": "Code",
                        "description": "A unique code associated with this specific type of error"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5,
                        "title": "Title",
                        "description": "A human-readable title for this error; displayable as an error dialog title"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "A human-readable error message, meant to be displayed to an end user or developer"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ServiceBaseModel"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "title",
                    "message"
                ],
                "title": "ErrorResponse",
                "description": "\n    A generic error object used for all error responses.\n\n    See [the error docs](docs/errors.md) for more information.\n    "
            },
...
            "ServiceBaseModel": {
                "properties": {},
                "type": "object",
                "title": "ServiceBaseModel"
            },
...
        }
    },
...
}